Repanic in waiting goroutine.
The default Go behavior for panics makes sense when goroutines aren't
related, but the whole point of this package is to create a hierarchical
tree of goroutines so that the lifetime of a goroutine created with 
gock is entirely contained in the caller goroutine's lifetime.

It then only makes sense that a panic is bubbled up not only in the
current goroutine but in its ancestors.