Remove optimization that breaks panic behavior.
As an optimization, we were running the first function provided to Wait
in the same goroutine that calls Wait. Due to this, we weren't calling
wait on the underlying bundle running the other functions until after
the first function is done, so, if one of these panicked, they went
unnoticed.

Now Wait is just a dumb helper for Bundle.